Update article text and add note on new Sqrt function
Updated the CodeProject article text to include a thank you note to the authors of the HTML to Markdown conversion tool. Added a note about the new version of the Sqrt function, called Newton Plus, which offers significant performance improvements for large numbers, while retaining the old version for smaller numbers.